Demo: Refactor demo app to reduce single component complexity.
Refactored the demo app Chat container component.
Everything was packed into this one container, so it made it confusing to understand the separation of features.
This is a first pass, for now I have just moved things into their own components, but we should look at ensuring each is fully isolated and reusable.
