+++ demo/src/components/ChatBoxComponent/index.ts
@@ -1,0 +1,1 @@
+export { ChatBoxComponent } from './ChatBoxComponent.tsx';
