+++ demo/src/components/ReactionComponent/index.ts
@@ -1,0 +1,1 @@
+export { ReactionComponent } from './ReactionComponent.tsx';
