+++ demo/src/components/TypingIndicatorPanel/index.ts
@@ -1,0 +1,1 @@
+export { TypingIndicatorPanel } from './TypingIndicatorPanel.tsx';
